:recycle: extract new module spring-boot-mongo
