+++ Spring-Boot/Spring-Boot-Problem/pom.xml
@@ -32,6 +32,10 @@
         </dependency>
         <dependency>
             <groupId>tw.waterball.judge-girl</groupId>
+            <artifactId>spring-boot-mongo</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>tw.waterball.judge-girl</groupId>
             <artifactId>judgegirl-problem</artifactId>
         </dependency>
         <dependency>
