+++ Spring-Boot/Spring-Boot-Submission/pom.xml
@@ -36,6 +36,10 @@
         </dependency>
         <dependency>
             <groupId>tw.waterball.judge-girl</groupId>
+            <artifactId>spring-boot-mongo</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>tw.waterball.judge-girl</groupId>
             <artifactId>spring-boot-amqp</artifactId>
         </dependency>
         <dependency>
