+++ Spring-Boot/pom.xml
@@ -35,6 +35,7 @@
         <module>Spring-Boot-Student</module>
         <module>Spring-Boot-Academy</module>
         <module>Spring-Boot-AMQP</module>
+        <module>Spring-Boot-Mongo</module>
         <module>Migration</module>
     </modules>
 
@@ -50,10 +51,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-mongodb</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
